refactor: Rename utils to scripts
