+++ scripts/templates.py
